divided Dropdown component + styling
