+++ components/Dropdown/Modal.tsx
@@ -1,9 +1,12 @@
-import React from "react";
-import { StyleSheet } from "react-native";
+import React from 'react';
+import { StyleSheet } from 'react-native';
 
-import Button from "./Button";
-import { View } from "./Themed";
+import Button from '../Button';
+import { View } from '../Themed';
 
+/**
+ * TODO: convert into General-use Modal, or delete
+ */
 interface ModalArgs {
   title?: string;
   onPressCancel?: ({ ...args }: any) => void;
